fix and and task 2
task 1
the error was in linking the library
fix

TASK 2
Do this task in 02-lightbox.html and 02-lightbox.js files. Break it down into several subtasks:

Create and render markup based on the galleryItems data array and the provided gallery item template. Use the ready code from the first task.
Connecting the script and styles of the library using the cdnjs CDN service. You need to add links to two files: simple-lightbox.min.js and simple-lightbox.min.css.
Library initialization after gallery elements are created and added to div.gallery. To do this, read the SimpleLightbox documentation - primarily the "Usage" and "Markup" sections.
Look in the documentation for the "Options" section and add the display of captions to images from the alt attribute. Let the caption be at the bottom and appear 250 milliseconds after the image is opened.

Выполняй это задание в файлах 02-lightbox.html и 02-lightbox.js. Разбей его на несколько подзадач:

Создание и рендер разметки по массиву данных galleryItems и предоставленному шаблону элемента галереи. Используй готовый код из первого задания.
Подключение скрипта и стилей библиотеки используя CDN сервис cdnjs. Необходимо добавить ссылки на два файла: simple-lightbox.min.js и simple-lightbox.min.css.
Инициализация библиотеки после того как элементы галереи созданы и добавлены в div.gallery. Для этого ознакомься с документацией SimpleLightbox - в первую очередь секции «Usage» и «Markup».
Посмотри в документации секцию «Options» и добавь отображение подписей к изображениям из атрибута alt. Пусть подпись будет снизу и появляется через 250 миллисекунд после открытия изображения.